[MOVE]/#6 interface 폴더 구조 수정
