--- src/interface/user/SocialUserInfo.ts
@@ -1,3 +1,0 @@
-export interface SocialUserInfo {
-    email: string
-}
