--- src/interface/user/UserCreateDto.ts
@@ -1,5 +1,0 @@
-export interface UserCreateDto {
-    email: string,
-    nickname: string,
-    profileImageId: string,
-}
