--- src/interface/user/UserInfo.ts
@@ -1,6 +1,0 @@
-export interface UserInfo {
-  email: string;
-  nickname: string;
-  profileImageId: string;
-  isDeleted: boolean;
-}
