--- src/interface/user/UserResponseDto.ts
@@ -1,9 +1,0 @@
-import mongoose from "mongoose";
-
-export interface UserResponseDto {
-    id: mongoose.Schema.Types.ObjectId,
-    nickname: string,
-    email: string,
-    profileImageId: string,
-    accessToken?: string,
-}
