--- src/interface/user/UserUpdateDto.ts
@@ -1,4 +1,0 @@
-export interface UserUpdateDto {
-    nickname: string,
-    profileImageId: string
-}
